+++ common/src/main/java/net/mca/client/gui/DestinyScreen.java
@@ -137,7 +137,7 @@ public class DestinyScreen extends VillagerEditorScreen {
                     int rows = (int)Math.ceil(Config.getInstance().destinyLocations.size() / 3.0f);
                     float offsetX = (y + 1) == rows ? (2 - (Config.getInstance().destinyLocations.size() - 1) % 3) / 2.0f : 0;
                     float offsetY = Math.max(0, 3 - rows) / 2.0f;
-                    addDrawableChild(new ButtonWidget((int)(width / 2 - 96 * 1.5f + (x + offsetX) * 96), (int)(height / 2 + (y + offsetY) * 20 - 16), 96, 20, new TranslatableText("gui.destiny." + location), sender -> {
+                    addDrawableChild(new ButtonWidget((int)(width / 2 - 96 * 1.5f + (x + offsetX) * 96), (int)(height / 2 + (y + offsetY) * 20 - 16), 96, 20, new TranslatableText("gui.destiny." + new Identifier(location).getPath()), sender -> {
                         selectStory(location);
                     }));
                     x++;
@@ -170,7 +170,7 @@ public class DestinyScreen extends VillagerEditorScreen {
         story.add(new TranslatableText("destiny.story.reason"));
         Map<String, String> map = Config.getInstance().destinyLocationsToTranslationMap;
         story.add(new TranslatableText(map.getOrDefault(location, map.getOrDefault("default", "missing_default"))));
-        story.add(new TranslatableText("destiny.story." + location));
+        story.add(new TranslatableText("destiny.story." + new Identifier(location).getPath()));
         this.location = location;
         setPage("story");
     }