+++ common/src/main/java/net/mca/entity/VillagerEntityMCA.java
@@ -76,10 +76,7 @@ import net.minecraft.stat.Stats;
 import net.minecraft.text.LiteralText;
 import net.minecraft.text.Text;
 import net.minecraft.text.TranslatableText;
-import net.minecraft.util.ActionResult;
-import net.minecraft.util.Formatting;
-import net.minecraft.util.Hand;
-import net.minecraft.util.Identifier;
+import net.minecraft.util.*;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Box;
 import net.minecraft.util.math.MathHelper;
@@ -96,10 +93,7 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.lang.reflect.Field;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.UUID;
+import java.util.*;
 import java.util.function.Predicate;
 
 import static net.mca.client.model.CommonVillagerModel.getVillager;
@@ -115,6 +109,8 @@ public class VillagerEntityMCA extends VillagerEntity implements VillagerLike<Vi
 
     private int despawnDelay;
     private int burned;
+
+    public final ConversationManager conversationManager = new ConversationManager(this);
 
     @Override
     public PlayerModel getPlayerModel() {
@@ -1028,7 +1024,7 @@ public class VillagerEntityMCA extends VillagerEntity implements VillagerLike<Vi
         } else if (Config.getInstance().useVanillaVoices) {
             return super.getDeathSound();
         } else {
-            return null;
+            return SoundsMCA.SILENT.get();
         }
     }
 
@@ -1075,11 +1071,11 @@ public class VillagerEntityMCA extends VillagerEntity implements VillagerLike<Vi
                 return getGenetics().getGender() == Gender.MALE ? mood.getSoundMale() : mood.getSoundFemale();
             }
 
-            return null;
+            return SoundsMCA.SILENT.get();
         } else if (Config.getInstance().useVanillaVoices) {
             return super.getAmbientSound();
         } else {
-            return null;
+            return SoundsMCA.SILENT.get();
         }
     }
 
@@ -1087,8 +1083,10 @@ public class VillagerEntityMCA extends VillagerEntity implements VillagerLike<Vi
     protected final SoundEvent getHurtSound(DamageSource cause) {
         if (Config.getInstance().useMCAVoices) {
             return getGenetics().getGender() == Gender.MALE ? SoundsMCA.VILLAGER_MALE_HURT.get() : SoundsMCA.VILLAGER_FEMALE_HURT.get();
-        } else {
+        } else if (Config.getInstance().useVanillaVoices) {
             return super.getHurtSound(cause);
+        } else {
+            return SoundsMCA.SILENT.get();
         }
     }